requirements.txt
streamlit
pandas
numpy
plotly
matplotlib
fpdf2
openai

